requirements.txt
blinker==1.8.2
click==8.1.7
flask==3.0.3
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
python-decouple==3.8
python-dotenv==1.0.1
werkzeug==3.0.6
zipp==3.20.2
